pyproject.toml
[tool.poetry]
name = "v3-stb-dot-com"
version = "0.1.0"
description = ""
authors = ["jacoacoacob <jacoacoacob@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
wagtail = "^4.0.1"
psycopg2 = "^2.9.3"
django-dotenv = "^1.4.2"
django-extensions = "^3.2.1"
gunicorn = "^20.1.0"
django-storages = "^1.13.1"
boto3 = "^1.24.71"
dj-database-url = "^1.0.0"
django-tailwind = "^3.4.0"
whitenoise = "^6.2.0"
django-debug-toolbar = "^3.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
